Extract handlers in separate function
