+++ packages/react-native-gesture-handler/src/v3/hooks/callbacks/js/useGestureStateChangeEvent.ts
@@ -1,15 +1,8 @@
-import { CallbackHandlers } from '../../../types';
+import { extractStateChangeHandlers } from '../../utils';
 import { onGestureHandlerStateChange } from '../onGestureHandlerStateChange';
 
 export function gestureStateChangeEvent(handlerTag: number, config: any) {
-  const { onBegin, onStart, onEnd, onFinalize } = config;
-
-  const handlers: CallbackHandlers = {
-    ...(onBegin ? { onBegin } : {}),
-    ...(onStart ? { onStart } : {}),
-    ...(onEnd ? { onEnd } : {}),
-    ...(onFinalize ? { onFinalize } : {}),
-  };
+  const handlers = extractStateChangeHandlers(config);
 
   return onGestureHandlerStateChange(handlerTag, handlers);
 }
