+++ packages/react-native-gesture-handler/src/v3/hooks/callbacks/reanimated/useReanimatedStateChangeEvent.ts
@@ -1,16 +1,9 @@
 import { Reanimated } from '../../../../handlers/gestures/reanimatedWrapper';
-import { CallbackHandlers } from '../../../types';
+import { extractStateChangeHandlers } from '../../utils';
 import { onGestureHandlerStateChange } from '../onGestureHandlerStateChange';
 
 export function useReanimatedStateChangeEvent(handlerTag: number, config: any) {
-  const { onBegin, onStart, onEnd, onFinalize } = config;
-
-  const handlers: CallbackHandlers = {
-    ...(onBegin ? { onBegin } : {}),
-    ...(onStart ? { onStart } : {}),
-    ...(onEnd ? { onEnd } : {}),
-    ...(onFinalize ? { onFinalize } : {}),
-  };
+  const handlers = extractStateChangeHandlers(config);
 
   const callback = onGestureHandlerStateChange(handlerTag, handlers);
