+++ packages/react-native-gesture-handler/src/v3/hooks/callbacks/reanimated/useReanimatedTouchEvent.ts
@@ -1,17 +1,9 @@
 import { Reanimated } from '../../../../handlers/gestures/reanimatedWrapper';
-import { CallbackHandlers } from '../../../types';
+import { extractTouchHandlers } from '../../utils';
 import { onGestureHandlerTouchEvent } from '../onGestureHandlerTouchEvent';
 
 export function useReanimatedTouchEvent(handlerTag: number, config: any) {
-  const { onTouchesDown, onTouchesMove, onTouchesUp, onTouchesCancelled } =
-    config;
-
-  const handlers: CallbackHandlers = {
-    ...(onTouchesDown ? { onTouchesDown } : {}),
-    ...(onTouchesMove ? { onTouchesMove } : {}),
-    ...(onTouchesUp ? { onTouchesUp } : {}),
-    ...(onTouchesCancelled ? { onTouchesCancelled } : {}),
-  };
+  const handlers = extractTouchHandlers(config);
 
   const callback = onGestureHandlerTouchEvent(handlerTag, handlers);
